% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate.theta.tcltk.r
\name{calculate.theta.tcltk}
\alias{calculate.theta.tcltk}
\title{Tcltk interfaces for \code{calculate.theta} function}
\usage{
calculate.theta.tcltk(input.matrix, species.data = NULL, juicer = T)
}
\arguments{
\item{input.matrix}{Matrix (or data.frame) with species composition data.}

\item{species.data}{Matrix (or data.frame) with species data. If suplied, it should have at least two columns - the first containing species name, the second containing layer.}

\item{juicer}{Argument specific for launching the function from JUICE software; logical (default = F) - is the function launched from JUICE? If \code{juicer = TRUE}, function is expecting that \code{species.data} have JUICE-specific structure, which enables to import data back to JUICE.}
}
\description{
Tcltk interfaces for \code{calculate.theta} function
}
\details{
Function \code{calculate.theta.tcltk} launches tcltk clickable interface, which enables to select methods and parameters used for calculation.
}
\author{
David Zeleny
}

