% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate.theta.R
\name{calculate.theta}
\alias{beals.2}
\alias{calculate.theta}
\alias{calculate.theta.0}
\title{Calculates species habitat specialization using co-occurrence based metrics}
\usage{
calculate.theta(input.matrix, species.data = NULL, reps = 10, psample = 5,
  method = "multiplicative", beals.file = NULL, parallel = F,
  no.cores = 2, remove.out = F, verbal = F, juicer = F, tcltk = F)

calculate.theta.0(input.matrix, species.data, sp, select.spp, remove.out,
  psample, reps, method, parallel, win.pb, verbal, juicer)

beals.2(x, include = TRUE, verbal = FALSE)
}
\description{
Calculates species habitat specialization using co-occurrence based metrics
}
\author{
David Zeleny (partly based on code written by Jason Fridley, Fridley et al. 2007)
}

